# opti: remove usage of modulo