--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,6 +1,12 @@
 #### Unreleased
 
 * Implement Percentage Price Oscillator (PPO)
+* More efficient BollingerBands
+* More efficient FastStochastic
+* More efficient SlowStochastic
+* More efficient StandardDeviation
+* More efficient Minimum
+* More efficient Maximum
 
 #### v0.2.0 - 2020-08-31