Remove usage of modulo to find the index of the next poped value for `Minimum`, `Maximum` and `StandardDeviation`.

Here is the benchmark with my i7 6500u:
before:
```
test Maximum                            ... bench:      75,284 ns/iter (+/- 4,688)
test Minimum                            ... bench:      77,725 ns/iter (+/- 10,071)
test StandardDeviation                  ... bench:      68,862 ns/iter (+/- 5,893)

test BollingerBands                     ... bench:      86,684 ns/iter (+/- 7,133)
test FastStochastic                     ... bench:     163,767 ns/iter (+/- 22,085)
test SlowStochastic                     ... bench:     178,034 ns/iter (+/- 12,981)
```
after:
```
test Maximum                            ... bench:      32,359 ns/iter (+/- 1,353)
test Minimum                            ... bench:      32,091 ns/iter (+/- 1,369)
test StandardDeviation                  ... bench:      36,202 ns/iter (+/- 1,989)

test BollingerBands                     ... bench:      41,592 ns/iter (+/- 1,635)
test FastStochastic                     ... bench:      72,559 ns/iter (+/- 1,773)
test SlowStochastic                     ... bench:      92,549 ns/iter (+/- 2,997)
```